# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: Semantic Business Map API
  version: 1.0.0
  description: Evidence-first application contract. Source for generated frontend types. No graph snippets are returned by the canvas endpoint.
servers:
  - url: /api/v1
paths:
...
  /projects/{projectId}/analysis-runs:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      operationId: listAnalysisRuns
      responses:
        '200':
          description: Analysis history
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/AnalysisRun' } }
    post:
      operationId: startAnalysis
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AnalysisInput' }
      responses:
        '201':
          description: Queued analysis run
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisRun' }
  /analysis-runs/{analysisRunId}:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getAnalysisRun
      responses:
        '200':
          description: Run state; progress is a percentage from 0 to 100
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AnalysisRun' }
  /analysis-runs/{analysisRunId}/cancel:
    parameters:
      - $ref: '#/components/parameters/RunId'
    post:
      operationId: cancelAnalysis
      responses:
        '200': { description: Cancellation accepted; client refreshes run state }
        '204': { description: Cancellation accepted }
  /analysis-runs/{analysisRunId}/events:
    parameters:
      - $ref: '#/components/parameters/RunId'
      - in: header
        name: Last-Event-ID
        schema: { type: string }
        description: Last received SSE id; events must have an id for automatic browser replay
    get:
      operationId: streamAnalysisEvents
      description: SSE messages may be unnamed or have event names equal to their data.type. Clients handle all event names, preserve Last-Event-ID on reconnect, and deduplicate by sequenceNumber.
      responses:
        '200':
          description: Server sent events, data JSON matches AnalysisEvent
          content:
            text/event-stream:
              schema: { type: string }
...
  /analysis-runs/{analysisRunId}/coverage:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getCoverage
      responses:
        '200':
          $ref: '#/components/responses/Inspection'
  /analysis-runs/{analysisRunId}/diagnostics:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getDiagnostics
      responses:
        '200':
          $ref: '#/components/responses/Inspection'
  /analysis-runs/{analysisRunId}/analyzer-plan:
    parameters:
      - $ref: '#/components/parameters/RunId'
    get:
      operationId: getAnalyzerPlan
      responses:
        '200':
          $ref: '#/components/responses/Inspection'
...
components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: string }
    RunId:
      in: path
      name: analysisRunId
      required: true
      schema: { type: string }
...
  responses:
    Inspection:
      description: Structured diagnostic data; fields vary by analyzer capability
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InspectionData'
...
  schemas:
...
    AnalysisInput:
      type: object
      properties:
        baselineRunId: { type: string, format: uuid }
        ref: { type: string }
    AnalysisRun:
      type: object
      required: [id, projectId, status, revision, createdAt, finishedAt, progress, currentStage]
      properties:
        id: { type: string }
        projectId: { type: string }
        status: { type: string }
        revision: { type: string, nullable: true }
        createdAt: { type: string }
        finishedAt: { type: string, nullable: true }
        progress: { type: number, minimum: 0, maximum: 100 }
        currentStage: { type: string, nullable: true }
...
    InspectionData:
      oneOf:
        - { type: object, additionalProperties: {} }
        - { type: array, items: {} }
...